chore: Move client.go to folder gnoclient. (Same functionality)
